// devtools/projects/protocol/src/lib/messages.d.ts
export interface DebugSignalGraphNode {
    id: string;
    kind: string;
    epoch: number;
    label?: string;
    preview: Descriptor;
    debuggable: boolean;
}
export interface DebugSignalGraphEdge {
    /**
     * Index of a signal node in the `nodes` array that is a consumer of the signal produced by the producer node.
     */
    consumer: number;
    /**
     * Index of a signal node in the `nodes` array that is a producer of the signal consumed by the consumer node.
     */
    producer: number;
}
/**
 * A debug representation of the signal graph.
 */
export interface DebugSignalGraph {
    nodes: DebugSignalGraphNode[];
    edges: DebugSignalGraphEdge[];
}
export interface SignalNodePosition {
    element: ElementPosition;
    signalId: string;
}
// ...
export interface DevToolsNode<DirType = DirectiveType, CmpType = ComponentType> {
    element: string;
    directives: DirType[];
    component: CmpType | null;
    children: DevToolsNode<DirType, CmpType>[];
    nativeElement?: Node;
    resolutionPath?: SerializedInjector[];
    hydration: HydrationStatus;
    defer: DeferInfo | null;
    onPush?: boolean;
}
export interface SerializedInjector {
    id: string;
    name: string;
    type: 'imported-module' | 'environment' | 'element' | 'null' | 'hidden';
    node?: DevToolsNode;
    providers?: number;
}
export interface SerializedProviderRecord {
    token: string;
    type: 'type' | 'existing' | 'class' | 'value' | 'factory' | 'multi';
    multi: boolean;
    isViewProvider: boolean;
    index?: number | number[];
}
// ...
export interface Descriptor {
    expandable: boolean;
    value?: any;
    editable: boolean;
    type: PropType;
    preview: string;
    containerType: ContainerType;
}
export interface DirectivesProperties {
    [name: string]: Properties;
}
// ...
export type DirectiveMetadata = AngularDirectiveMetadata | AcxDirectiveMetadata | WizComponentMetadata;
export interface SerializedInjectedService {
    token: string;
    value: string;
    position: number[];
    flags?: InjectOptions;
    resolutionPath?: SerializedInjector[];
}
export interface Properties {
    props: {
        [name: string]: Descriptor;
    };
    metadata?: DirectiveMetadata;
}
export type ElementPosition = number[];
export interface DirectivePosition {
    element: ElementPosition;
    directive?: number;
}
// ...
export type PropertyQuery = AllPropertiesQuery | SelectedPropertiesQuery;
export interface ComponentExplorerViewQuery {
    selectedElement: ElementPosition;
    propertyQuery: PropertyQuery;
}
export interface ComponentExplorerView {
    forest: DevToolsNode[];
    properties?: DirectivesProperties;
}
// ...
export interface ElementProfile {
    directives: DirectiveProfile[];
    children: ElementProfile[];
    type: 'defer' | 'element';
}
export interface ProfilerFrame {
    source: string;
    duration: number;
    directives: ElementProfile[];
}
export interface UpdatedStateData {
    directiveId: DirectivePosition;
    keyPath: string[];
    newValue: any;
}
export interface Route {
    name?: string;
    hash?: string;
    specificity?: string;
    handler?: string;
    pathMatch?: 'prefix' | 'full';
    canActivateGuards?: string[];
    canActivateChildGuards?: string[];
    canMatchGuards?: string[];
    canDeactivateGuards?: string[];
    providers?: string[];
    title?: string;
    children?: Array<Route>;
    data?: any;
    path: string;
    component: string;
    isActive: boolean;
    isAux: boolean;
    isLazy: boolean;
    isRedirect: boolean;
}
export interface AngularDetection {
    isAngularDevTools: true;
    isIvy: boolean;
    isAngular: boolean;
    isDebugMode: boolean;
    isSupportedAngularVersion: boolean;
}
export type Topic = keyof Events;
export interface SupportedApis {
    profiler: boolean;
    dependencyInjection: boolean;
    routes: boolean;
    signals: boolean;
    transferState: boolean;
    signalPropertiesInspection: boolean;
}
export type TransferStateValue = string | number | boolean | null | undefined | Record<string, unknown> | unknown[];
export interface Events {
    handshake: () => void;
    shutdown: () => void;
    queryNgAvailability: () => void;
    ngAvailability: (config: {
        version: string | undefined;
        devMode: boolean;
        ivy: boolean;
        hydration: boolean;
        supportedApis: SupportedApis;
    }) => void;
    inspectorStart: () => void;
    inspectorEnd: () => void;
    getSignalGraph: (query: ElementPosition) => void;
    latestSignalGraph: (graph: DebugSignalGraph) => void;
    getSignalNestedProperties: (position: SignalNodePosition, path: string[]) => void;
    signalNestedProperties: (position: SignalNodePosition, data: Properties, path: string[]) => void;
    getNestedProperties: (position: DirectivePosition, path: string[]) => void;
    nestedProperties: (position: DirectivePosition, data: Properties, path: string[]) => void;
    setSelectedComponent: (position: ElementPosition) => void;
    getRoutes: () => void;
    updateRouterTree: (routes: Route[]) => void;
    navigateRoute: (route: string) => void;
    componentTreeDirty: () => void;
    getLatestComponentExplorerView: (query?: ComponentExplorerViewQuery) => void;
    latestComponentExplorerView: (view: ComponentExplorerView) => void;
    updateState: (value: UpdatedStateData) => void;
    startProfiling: () => void;
    stopProfiling: () => void;
    sendProfilerChunk: (results: ProfilerFrame) => void;
    profilerResults: (results: ProfilerFrame) => void;
    createHighlightOverlay: (position: ElementPosition) => void;
    removeHighlightOverlay: () => void;
    createHydrationOverlay: () => void;
    removeHydrationOverlay: () => void;
    highlightComponent: (id: number) => void;
    selectComponent: (id: number) => void;
    removeComponentHighlight: () => void;
    enableTimingAPI: () => void;
    disableTimingAPI: () => void;
    getInjectorProviders: (injector: SerializedInjector) => void;
    latestInjectorProviders: (injector: SerializedInjector, providers: SerializedProviderRecord[]) => void;
    logProvider: (injector: SerializedInjector, providers: SerializedProviderRecord) => void;
    getTransferState: () => void;
    transferStateData: (data: Record<string, TransferStateValue> | null) => void;
    contentScriptConnected: (frameId: number, name: string, url: string) => void;
    contentScriptDisconnected: (frameId: number, name: string, url: string) => void;
    enableFrameConnection: (frameId: number, tabId: number) => void;
    frameConnected: (frameId: number) => void;
    detectAngular: (detectionResult: AngularDetection) => void;
    backendReady: () => void;
    log: (logEvent: {
        message: string;
        level: 'log' | 'warn' | 'debug' | 'error';
    }) => void;
}